+++ my-app/Frontend/src/Components/Profile/CreatorSubset/CreatorHome.tsx
@@ -1,12 +1,17 @@
 import React, { useEffect, useState } from "react";
-import CreatorList from "./CreatorList";
+import CreatorList from "./Home/CreatorList";
+import CreatorHomePosts from "./Home/CreatorHomePosts";
 import { getSubs } from "../../Posts/GetSubs";
 import { getUserProfile } from "../../../Pages/Profile/GetProfile";
 import { UserDataProps as UserData } from "../../../Types/ProfileNavbar";
+import { getPosts as getPostsAPI } from "../../Posts/GetPosts";
+import { Post } from "../../Posts/GetPosts";
 import { useParams } from "react-router-dom";
 
 const CreatorHome: React.FC = () => {
   const [subscribedUsers, setSubscribedUsers] = useState<UserData[]>([]);
+  const [creatorPosts, setCreatorPosts] = useState<Post[]>([]); 
+
   const { userId } = useParams(); 
 
   useEffect(() => {
@@ -29,7 +34,19 @@ const CreatorHome: React.FC = () => {
       }
     };
 
+    const fetchCreatorPosts = async () => {
+      if (!userId) return;
+
+      try {
+        const posts = await getPostsAPI(userId);
+        setCreatorPosts(posts);
+      } catch (error) {
+        console.error("Error fetching creator posts:", error);
+      }
+    };
+
     fetchFollowing();
+    fetchCreatorPosts();
   }, [userId]);
 
   return (
@@ -39,8 +56,7 @@ const CreatorHome: React.FC = () => {
       </div>
 
       <div>
-        <h1 className="profile_subTitle">Posts</h1>    
-       
+      <CreatorHomePosts posts={creatorPosts} />
       </div>
     </>
   );
